requirements.txt
openmetadata-ingestion==1.1.5